Fix numeric overflow

+++ db/migrate/20141223044623_convert_weights_to_decimals.rb
@@ -1,12 +1,12 @@
 class ConvertWeightsToDecimals < ActiveRecord::Migration
   def change
-    change_column :bags, :current_weight, :decimal, precision: 10, scale: 4
-    change_column :bags, :initial_weight, :decimal, precision: 10, scale: 4
+    change_column :bags, :current_weight, :decimal, precision: 16, scale: 4
+    change_column :bags, :initial_weight, :decimal, precision: 16, scale: 4
 
-    change_column :jars, :current_weight, :decimal, precision: 10, scale: 4
-    change_column :jars, :initial_weight, :decimal, precision: 10, scale: 4
+    change_column :jars, :current_weight, :decimal, precision: 16, scale: 4
+    change_column :jars, :initial_weight, :decimal, precision: 16, scale: 4
 
-    change_column :lots, :current_weight, :decimal, precision: 10, scale: 4
-    change_column :lots, :initial_weight, :decimal, precision: 10, scale: 4
+    change_column :lots, :current_weight, :decimal, precision: 16, scale: 4
+    change_column :lots, :initial_weight, :decimal, precision: 16, scale: 4
   end
 end
